+++ packages/implementations/react/src/composite/Datepicker/src/components/Footer/Footer.tsx
@@ -1,5 +1,8 @@
 import React from "react";
-import { CalendarEventOutlinedIcon } from "@rarui/icons";
+import {
+  CalendarEventOutlinedIcon,
+  CancelCircleOutlinedIcon,
+} from "@rarui/icons";
 import { Box } from "@rarui-react/box";
 import { Icon } from "@rarui-react/icon";
 import { Button } from "@rarui-react/button";
@@ -10,6 +13,7 @@ const DatepickerFooter: React.FC<DatepickerFooterProps> = ({
   className: _className,
   style: _style,
   setSelected,
+  clear,
   ...props
 }) => {
   const dataTestid = (props as { "data-testid": string })?.["data-testid"];
@@ -17,13 +21,14 @@ const DatepickerFooter: React.FC<DatepickerFooterProps> = ({
   return (
     <div style={{ clear: "both" }}>
       <Box
-        justifyContent="center"
+        justifyContent="space-between"
         borderWidth="$none"
         borderColor="$divider"
         borderTopWidth="$1"
         borderStyle="solid"
         py="$2xs"
-        px="$xs"
+        px="$s"
+        display="flex"
       >
         <Button
           type="button"
@@ -35,6 +40,17 @@ const DatepickerFooter: React.FC<DatepickerFooterProps> = ({
         >
           <Icon source={<CalendarEventOutlinedIcon />} />
           Hoje
+        </Button>
+        <Button
+          type="button"
+          size="small"
+          variant="tonal"
+          onClick={clear}
+          data-testid={dataTestid ?? "clear-button"}
+          {...props}
+        >
+          <Icon source={<CancelCircleOutlinedIcon />} />
+          Limpar
         </Button>
       </Box>
     </div>