+++ packages/implementations/react/src/composite/Datepicker/src/components/Footer/footer.spec.tsx
@@ -5,7 +5,7 @@ import { DatepickerFooter } from "./Footer";
 import { DatepickerFooterProps } from "./footer.types";
 
 const makeSut = (props: DatepickerFooterProps) => {
-  render(<DatepickerFooter {...props} data-testid="today-button" />);
+  render(<DatepickerFooter {...props} />);
 };
 
 describe("GIVEN <Datepicker.Footer />", () => {
@@ -15,6 +15,12 @@ describe("GIVEN <Datepicker.Footer />", () => {
       makeSut({ setSelected: mockSetSelected });
       fireEvent.click(screen.getByTestId("today-button"));
       expect(mockSetSelected).toHaveBeenCalled();
+    });
+    it("THEN should call the clear function by clicking the clear button", () => {
+      const mockClear = jest.fn();
+      makeSut({ clear: mockClear });
+      fireEvent.click(screen.getByTestId("clear-button"));
+      expect(mockClear).toHaveBeenCalled();
     });
   });
 });