+++ packages/implementations/react/src/composite/Tabs/src/tabs.types.ts
@@ -18,6 +18,12 @@ export interface TabsTyping {
    * If unset, defaults to the first tab. This prop should be a number representing the index of the tab.
    */
   preSelectedTab?: number;
+
+  /**
+   * Sets the indicated tab as current active tab.
+   * Can be used to change the active tab from outside of the component
+   */
+  selectedTab?: number;
 }
 
 export type TabsProperties = TabsTyping &